+++ src/python/zensols/util/__init__.py
@@ -4,12 +4,51 @@ and fork/exec operations.
 """
 __author__ = 'Paul Landes'
 
+import sys
+from typing import Any
+from dataclasses import dataclass, field
+import traceback
+
 
 class APIError(Exception):
     """Base exception from which almost all library raised errors extend.
 
     """
     pass
+
+
+@dataclass
+class Failure(object):
+    """Contains information for failures as caught exceptions used by
+    programatic methods.
+
+    """
+    exception: Exception = field()
+    """The exception that was generated."""
+
+    thrower: Any = field(default=None)
+    """The insstance of the class that is throwing the exception."""
+
+    traceback: traceback = field(default=None, repr=False)
+    """The stack trace."""
+
+    message: str = field(default=None)
+    """A high level explaination of what went wrong"""
+
+    def __post_init__(self):
+        self._exec_info = sys.exc_info()
+        if self.traceback is None:
+            self.traceback = self._exec_info[2]
+
+    def print_stack(self):
+        traceback.print_exception(*self._exec_info)
+
+    def __str__(self) -> str:
+        msg: str = str(self.exception) if self.message is None else self.message
+        return str(f'{type(self.exception).__name__}: {msg}')
+
+    def __repr__(self) -> str:
+        return self.__str__()
 
 
 from .std import *